refactor: 게임 결과인 GameResult 클래스를 도메인 패키지로 옮겼다.
사용자에게 결과값을 출력한다는 요구사항을 충족시키기 위해서 필수적인 클래스이다.
